# go-server-server/go/model_teacher.go
/*
 * Swagger Studentcourse
 *
 * This is a sample server Studentcourse server.  You can find out more about     Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).      For this sample, you can use the api key `special-key` to test the authorization     filters.
 *
 * API version: 1.0.0
 * Contact: apiteam@swagger.io
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type Teacher struct {
	Id int64 `json:"id,omitempty"`

	Teachername string `json:"teachername,omitempty"`

	PopularityScore float64 `json:"popularityScore,omitempty"`

	Courses []interface{} `json:"courses,omitempty"`
}